Add acceptance tests for auth-switch
This is the trickiest command of them all, requiring us to log in using 2
sets of credentials. Thus, I've implemented some optional env vars that can
be specified when a command is run to make this one work:
GH_ACCEPTANCE_USER, GH_ACCEPTANCE_HOST_ALT, GH_ACCEPTANCE_USER_ALT, and
GH_ACCEPTANCE_TOKEN_ALT, and GH_ACCEPTANCE_TOKEN_ALT. All of the ALT
variables correspond to a second user account you are currently logged in
to for testing.

With these additional variables, the test logs in to both instances and
then tests that gh auth switch works to go between the two. I think the
test is rather straight forward if the setup is not.